Refactored the DAO folder with a DAO for teacher and student and have implimented the classroom view but there is still work that needs to be done
