+++ src/main/java/com/example/cab302tailproject/DAO/UserDAO.java
@@ -1,0 +1,6 @@
+package com.example.cab302tailproject.DAO;
+
+public interface UserDAO {
+    boolean checkEmail(String email);
+    boolean checkPassword(String email, String password);
+}
