+++ src/main/java/com/example/cab302tailproject/controller/teachercontroller/AiAssistantController.java
@@ -94,7 +94,24 @@ public class AiAssistantController {
     // --- Event Handlers for Top Navigation Buttons ---
     // TODO (Placeholder implementations - add actual navigation/logic)
     @FXML private void onFilesClicked(ActionEvent event) { System.out.println("Files button clicked."); }
-    @FXML private void onStudentsClicked(ActionEvent event) { System.out.println("Students button clicked."); }
+
+    @FXML
+    private void onStudentsClicked(ActionEvent event) throws IOException {
+        // Load the FXML
+        FXMLLoader fxmlLoader = new FXMLLoader(TailApplication.class.getResource("classroom-teacher-view.fxml"));
+        Scene scene = new Scene(fxmlLoader.load(), TailApplication.WIDTH, TailApplication.HEIGHT);
+
+        // Get the controller associated with the FXML
+        ClassroomViewController controller = fxmlLoader.getController();
+
+        // Now call the method on the actual instance
+        controller.loadStudentData();
+
+        // Set the new scene
+        Stage stage = (Stage) studentsButton.getScene().getWindow();
+        stage.setScene(scene);
+    }
+
     @FXML private void onHomeClicked(ActionEvent event) { System.out.println("Home button clicked."); }
     @FXML private void onSettingsClicked(ActionEvent event) { System.out.println("Settings button clicked."); }
